can't index objects and image needs restructuring
