+++ src/main/js/Images.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import {Container} from "react-bootstrap";
+import {Container, Col, Row, Image} from "react-bootstrap";
+
 import { planeQuestion } from '../../test/resources/ExampleQuestions';
 import img from '../../main/js/images/3CTransverse.jpg';
 
@@ -10,17 +11,21 @@ export const Images = props => {
     //const images = props.questionModel.imageUrl.map(displayImg => (<option key={displayImg}/>));
 
 
+    //display image seperately
     
 
     return (
         
-        <Container>
+       
         
-        <img src = {planeQuestion.imageUrl} alt="DID NOT WORK"/>
+        <Image img src = {planeQuestion.imageUrl} alt = "DIDNT WORK" style={{ 
+            width: 500, 
+            height: 400,
+        }}
+        />
         
-         
-        
-        </Container>
+      
+    
     );
 };
