+++ src/test/js/QuestionDemo.js
@@ -1,15 +1,18 @@
 import React from 'react';
 
 import {planeQuestion, structureQuestion} from "../resources/ExampleQuestions"
-import {Question} from "../../main/js/QuestionDropdown"
+import {Question} from "../../main/js/QuestionModel"
 import Images from "../../main/js/Images";
 
 export const QuestionDemo = props => {
     return (
         <div>
-            <Images questionModel={planeQuestion.imageUrl} />
+
+            <Images className="left" questionModel={planeQuestion.imageUrl}/>
+            
             <Question questionModel={planeQuestion} />
-            <Question questionModel={structureQuestion} />
+
+            
         </div>
     )
 }
